Consolidated invocation page functionalities
